refatorando funcoes e comeco da pep8
