+++ utils/cli.py
@@ -1,0 +1,18 @@
+# coding: utf-8
+import os
+
+
+def banner():
+
+        os.system('clear')
+        print("\n")
+        print("\033[32m\tMMP\"\"MM\"\"YMM `7MMF'     A     `7MF' db\"")
+        print("\033[32m\tP'   MM   `7   `MA     ,MA     ,V  ;MM:     ")
+        print("\033[33m\t     MM         VM:   ,VVM:   ,V  ,V^MM.    ")
+        print("\033[33m\t     MM          MM.  M' MM.  M' ,M  `MM    ")
+        print("\033[33m\t     MM          `MM A'  `MM A'  AbmmmqMA   ")
+        print("\033[31m\t     MM           :MM;    :MM;  A'     VML  ")
+        print("\033[31m\t   .JMML.          VF      VF .AMA.   .AMMA.\033[39m")
+        print("\t\tTWA Corp. sub domain finder")
+        print("\t           Use with NO moderation :D")
+        print("\t             Third World Attacker\n")
